refactor: Move stack to generic components folder
